ai-engine/requirements.txt
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.3
pdf2image>=1.16.3
pdfplumber>=0.10.2
python-docx>=1.1.0
pinecone-client>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.1.4
scikit-learn>=1.3.2
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.2
typing-extensions>=4.8.0
python-dotenv>=1.0.0
